+++ pkg/bwagreement/common.go
